@@ -15,15 +19,15 @@ android {
             storePassword 'caronte'
         }
     }
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.2'
     defaultConfig {
         applicationId "com.glm.trainer"
-        minSdkVersion 11
-        targetSdkVersion 23
+        minSdkVersion 23
+        targetSdkVersion 29
 
-        versionCode 143
-        versionName "8.1"
+        versionCode 176
+        versionName "8.6.0"
         archivesBaseName = "JustRunPRO-$versionName"
     }
     buildTypes {
