@@ -3,8 +3,8 @@ package com.glm.app.fragment;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.Toolbar;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.widget.Toolbar;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
@@ -29,6 +29,9 @@ import com.glm.trainer.R;
 import com.glm.utils.ExerciseUtils;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdView;
+import com.google.android.gms.ads.MobileAds;
+import com.google.android.gms.ads.initialization.InitializationStatus;
+import com.google.android.gms.ads.initialization.OnInitializationCompleteListener;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -102,6 +105,14 @@ public class SummaryFragment extends Fragment {
 
 		rootView = inflater.inflate(R.layout.new_summary_history,
 				container, false);
+
+		MobileAds.initialize(mContext, new OnInitializationCompleteListener() {
+			@Override
+			public void onInitializationComplete(InitializationStatus initializationStatus) {
+			}
+		});
+
+
 
 		mTotalGraph		= (RelativeLayout) rootView.findViewById(R.id.totalGraph);
 		mRunGraph		= (RelativeLayout) rootView.findViewById(R.id.runGraph);
